\name{basque}
 \docType{data}
 \alias{basque}
 \title{Wide-format Panel Data from Spanish Regions Used For Testing}
 \description{The dataset contains per-capita GDP from 1955--1997
on 17 Spanish regions, employing the terrorist conflict in the Basque Country
as an intervention to exploit the causal evidence of conflict on the economy.}

\usage{basque}

\format{A wide-format panel dataframe made up of 18 units: 1 treated (col no. 4; the Basque country) and 16 control regions (col no. 2-14, 6-19). The outcome is per-capita GDP and 42 time periods (1955-1997) are contained in the dataframe.}

\source{Abadie, A. and Gardeazabal, J. (2003) Economic Costs of Conflict: A Case Study of the Basque Country \emph{American Economic Review} 93 (1) 113--132.

Abadie, A., Diamond, A., Hainmueller, J. (2011). Synth: An R Package for Synthetic Control Methods in Comparative Case Studies. \emph{Journal of Statistical Software} 42 (13) 1--17.

}

\keyword{datasets}
