% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsam.R
\name{subsam}
\alias{subsam}
\title{Conduct subsampling method to produce statistical inference for average treatment effects estimated by synthetic method.}
\usage{
subsam(
  synth.meth.obj = NULL,
  m = NULL,
  type = "subsam",
  nb = 1000,
  conf = 0.95,
  opt.weights = NULL,
  Y1 = NULL,
  X1 = NULL,
  X2 = NULL,
  time.post = NULL,
  ATT_SC = NULL,
  sigma2_v = NULL
)
}
\arguments{
\item{synth.meth.obj}{The object that comes from running synth.meth(). This objects contains all matrices and parameters necessary for running the function subsam(). Therefore, once users import this object, there is no need to specify any of matrices and parameters manually.}

\item{m}{The value of subsample size. Noted that the value of m should not be larger than the number of pre-treatment periods.}

\item{type}{A character string indicating the type of applied inference method. Possible values are "subsam" (the default, subsampling method) and "par.boot" (parametric bootstrap method).}

\item{nb}{The number of subsampling-bootstrap replication.}

\item{conf}{A scalar or vector containing the confidence level(s) of the required interval(s)}

\item{Y1}{A matrix that contains values of the treatment unit over the pre-intervention periods.}

\item{X1}{A matrix that contains values of the control units over the pre-intervention periods.}

\item{X2}{A matrix that contains values of the control units over the post-intervention periods.}

\item{time.post}{The number of post-treatment periods.}

\item{ATT_SC}{The value of average treatment effect estimated by synthetic control method.}

\item{sigma2_v}{The variance of the estimated ATT over post-treatment periods.}

\item{opt.weight}{A Vector or matrix of the set of weights that generate the best fitting combination of the control units used for synthetic control.}

\item{Y2}{A matrix that contains values of the treatment unit over the post-intervention periods.}
}
\value{
A list which contains a series of statistics derived from the inference method and other components.
\item{ci_set}{A matrix of intervals across all the confidence level(s) required.}
\item{cr_mi}{The value of the least lower bound of intervals.}
\item{cr_max}{The value of the greatest upper bound of intervals.}
\item{ATT_order}{A matrix that contains a set of subsampling ATT's sorted by ascending order.}
\item{ATT_SC}{The value of estimated average treatment effect on treated.}
\item{nb}{The number of subsampling-bootstrap replication.}
}
\description{
This function provides users with formal inference methods that can be used to derive confidence intervals for the SC or MSC ATT estimators and conduct inference whereas the standard bootstrap cannot.
}
\details{
To run this function, we need users to supply data object obtained by running the synth.meth(), or users should manually input relevant matrices and numeric values to fill in the corresponding arguments. In this function, we provide two methods, including subsampling and parametric bootstrap, for statistical inference. Users can employ the argument "type" to specify which type of procedure to use.
}
\examples{
  ##In order to conduct inference properly, we recommended users to directly use the list object
  ##derived from running synth.meth() and choose appropriate subsample size m.

  #load data from the package.
  data(synth.data)

  #extract relevant components necessary for running synth.meth()
  #from wide-format panel data.
  synth.prep.out =
    synth.prep(
     data = synth.data,
     treatment.identifier = 9,
     controls.identifier = c(2:8),
     time.variable = 1,
     time.predictors.prior = c(1984:1989)
     )
  #then run the synth.meth command to identify the optimized weights estimated by SC method.
  synth.meth.out = synth.meth(synth.prep.out)

  #directly use this object to conduct formal inference process and then
  #obtain confidence intervals in different levels.
  subsam.out = subsam(synth.meth.out, m = 4, conf = c(0.99, 0.95, 0.90))
}
\references{
Li K, Shankar V (2020) Estimating the causal effect of a digitally native retailer opening a new store: a new two-step synthetic control method. Working Paper, University of Texas at Austin, Austin, TX.

  Li K, Shankar V (2020) Statistical inference for average treatment effects estimated by synthetic control methods. \emph{Journal of American Statistical Association}.
}
