\name{synth.data}
 \docType{data}
 \alias{synth.data}
 \title{Wide-format Panel Data Used For Testing}
 \description{This artificial panel data set is used to test the validity of corresponding inference method.
          }
\usage{synth.data}
\format{A dataframe made up of 8 units: 1 treated (col no. 9) and 7 control (col no. 2-8), toy production as the outcome,  21 time periods (1980 - 2000). All columns have column names.}
\source{Abadie, A. and Gardeazabal, J. (2003) Economic Costs of Conflict: A Case Study of the Basque Country \emph{American Economic Review} 93 (1) 113--132.

Abadie, A., Diamond, A., Hainmueller, J. (2011). Synth: An R Package for Synthetic Control Methods in Comparative Case Studies. \emph{Journal of Statistical Software} 42 (13) 1--17.

}
\keyword{datasets}
