% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{synth.meth}
\alias{synth.meth}
\title{Implement synthetic control method to estimate the average treatment effect.}
\usage{
synth.meth(
  synth.prep.obj = NULL,
  meth = "ordinary",
  inter = FALSE,
  Y1 = NULL,
  X1 = NULL,
  Y2 = NULL,
  X2 = NULL,
  time.variable = NULL
)
}
\arguments{
\item{meth}{A character string indicating the type of applied method. Possible values are "ordinary" (the default, SC method) and "modified" (MSC method).}

\item{inter}{A Logical flag. If true, an intercept term will be added as an independent
variable. By default, there is no space for the intercept term when applying SC method.
Whereas, the intercept term is required when choosing MSC method(No need to declare).}

\item{Y1}{A matrix of treatment unit data for pre-treatment periods.}

\item{X1}{A matrix of control units data for pre-treatment periods.}

\item{Y2}{A matrix of treatment unit data for post-treatment periods.}

\item{X2}{A matrix of control units data for post-treatment periods.}

\item{synth.prep.object}{The object that comes from running synth.prep(). This objects contains all information about four necessary matrices. Therefore, once users import this object, there is no need to specify any of four matrices manually.}

\item{time.overall}{A numeric vector identifying the row numbers corresponding to overall periods}
}
\value{
A list that contains the weights on control units and other components necessary for running \code{subsam}.
\item{b_SC}{A vector of weights across the controls.}
\item{ATT_S}{The value of estimated average treatment effect.}
\item{sigma_SC}{MSPE from optimization over weights on control units.}
\item{sigma2_v}{The variance of estimated average treatment effect.}
\item{r2_SC}{The value of R-squared.}
\item{r2_bar_SC}{The value of adjusted R-squared.}
\item{dataset}{A list that contains a series of matrices and scalars for future use.}
}
\description{
This function can routinely search for the set of weights that generate the best fitting combination of the control units chosen to approximate counterfactual outcomes of treatment unit and thus estimate SC ATE with long panels under general conditions. This function require the users to input four matrices as its main arguments. they are named Y1, X1, Y2, and X2 accordingly. Y1 and Y2 contain values of the treatment unit over the pre-intervention and post-intervention periods respectively. X1 and X2 contain the the values for the control units over the pre-intervention and post-intervention periods respectively.
}
\details{
Creating relevant matrices to be loaded into synth.meth() can be tedious. Users are recommended to apply a preparatory function called "synth.prep" that allows the user to easily create all inputs required for synth.meth(). For more information about the synthetic control method, please refer to the paper Statistical Inference for Average Effects Estimated by Synthetic Control Methods (Kathleen T. Li, 2019).
}
\examples{
  ##In order to run synth.meth() properly, we recommended users to run
  ##synth.prep for component-extraction first.

  #load data from the package.
  data(synth.data)

  #extract relevant components necessary for running synth.meth()
  #from wide-format panel data.
  synth.prep.out =
    synth.prep(
     data = synth.data,
     treatment.identifier = 9,
     controls.identifier = c(2:8),
     time.variable = 1,
     time.predictors.prior = c(1984:1989)
     )
  #now run the synth.meth command to identify the optimized weights
  synth.meth.out = synth.meth(synth.prep.out)

  #or users can choose MSC as estimation method.
  synth.meth.out2 = synth.meth(synth.prep.out, meth = "modified")
}
\references{
Li K, Shankar V (2020) Estimating the causal effect of a digitally native retailer opening a new store: a new two-step synthetic control method. Working Paper, University of Texas at Austin, Austin, TX.

  Li K, Shankar V (2020) Statistical inference for average treatment effects estimated by synthetic control methods. \emph{Journal of American Statistical Association}.
}
